Harden rstest helper collection review follow-ups
Address a further round of review feedback on rstest helper call-site
collection (8.2.2).

- Resolve method-call callees through the call expression's own HIR owner.
  The pass-owned `CallSiteVisitor` walks nested bodies without refreshing the
  ambient typeck context, so a method call inside a nested closure carried a
  different owner than `cx.typeck_results()`; reading `type_dependent_def_id`
  from the mismatched results tripped `validate_hir_id_for_typeck_results` and
  ICE'd the driver. `resolve_method_call_def_id` now mirrors
  `LateContext::qpath_res`: prefer owner-matching typeck results, otherwise
  load the owner's own, and yield `None` when none exist. A strict `#[rstest]`
  regression in `collection_zero_diagnostic` exercises a method call inside a
  closure and the harness summary now records nine call sites.
- Write the test-only collection summary through `cap_std::fs_utf8` instead of
  `std::fs::OpenOptions`, per the repository's capability-oriented filesystem
  policy. The harness supplies an absolute path, so the parent directory is
  opened with ambient authority and the summary appended relative to it.
- Split the example-harness lock protocol out of `tests/ui.rs` into
  `tests/ui/harness_lock.rs` (path-based module), keeping both files under the
  400-line limit and leaving `ui.rs` focused on the example and trybuild
  assertions.
