+++ Cargo.toml
@@ -14,6 +14,7 @@ whitaker-common = { path = "common", version = "0.2.6" }
 whitaker-installer = { path = "installer", version = "0.2.6" }
 whitaker = { path = ".", version = "0.2.6" }
 camino = "1.2.1"
+cap-std = { version = "3.4.5", features = ["fs_utf8"] }
 cargo_metadata = "0.23.0"
 clap = "4.5.38"
 directories-next = "2.0.0"
